refactor(macos): extract playlist domain from AppModel into AppModel+Playlists.swift
Phase 3 of the AppModel god-object split. Relocates the cohesive playlist
domain — actions (play/shuffle/playNext/addToQueue/toggleFavorite/
enqueueDownload), sidebar CRUD (create/rename/delete/duplicate, inline-edit
commit, drag-reorder), and sharing (web URL, copy link, open-in-Jellyfin,
M3U/JSON export) — out of the 7,106-line AppModel.swift into a new
`extension AppModel` file, mirroring the AppModel+Downloads.swift pattern.

Pure relocation: the 590 moved lines are byte-identical to what was removed
(no behaviour, signatures, or logic changed). AppModel.swift drops from 7,106
to 6,515 lines.

Mechanics:
- All `@Observable` stored state the methods read (playlists, playlistTracks,
  currentPlaylistTracks, playlistDescriptions, sidebarEditingPlaylistId,
  sidebarEditingDraft, sidebarCopyingPlaylistIds, playlistPendingDelete) stays
  declared on the main class — stored properties can't live in an extension.
- The cluster's two private helpers (resolveExportApiKey, writeExport) are
  self-contained, so they move with the domain and stay private.
- Single access-control promotion: setFavorite (private -> internal), the
  shared favorite primitive the moved toggleFavorite(playlist:) routes through
  (album/track/artist toggles call it too). Comment updated to reflect why.

Verified: clean rebuild (rm -rf macos/.build && swift build) and the full
swift test suite (889 tests, 0 failures) both green.
